add basic SDL2 backend
Add a basic SDL2 sys backend, that is not compiled by default. Currently
it only uses the raster renderer and blits it onto a texture that is
displayed.

Eventually it would be nice to enrich it with OpenGL rendering and imgui
in order to display debugging tools.

It's basic alright, but actually works pretty nice and may become the
default.