gfx: render game during blitframebuffer op execution
The current approach was to render the game after the VM is done
running, which introduced problems if e.g. a set_palette op was met
after the blitframebuffer op in the same cycle. For such cases the frame
would be rendered with an invalid palette.

Fix this by making the renderer draw the framebuffer when processing
blitframebuffer. This requires one bridge implementation of
gfx::Backend for all renderers, but also simplifies the game loop since
we don't need to care about asking the renderer to render the
framebuffer there anymore.