sys/sdl2: add keypress cooldown after gaining focus
Keys that are pressed as the window enters focus will be signaled as new
keypresses, which is not necessarily what we want. For instance, if the
window is focused using a keyboard shortcut that involves "esc", the
"esc" keypress will be registered and the game will exit.

Prevent this from happening by adding a 1-tick cooldown cycle where
keypresses are ignored after the window gains focus.