gfx: pass zoom factor of polygons to backend
Performing zoom in the VM means that each vertex of a polygon is in
(320,200) coordinates, which results in loss of precision if we draw at
a higher resolution.

Pass the zoom factor down to the backend and let it decide how zooming
should be done. This has the nice effect of letting us do the zoom as
part of the vertex shader when GL rendering is used.