gfx: pass coordinates to zoom as an extra parameter to the backend
Offsets in polygon hierarchies need to be zoomed, but not the position
of the polys themselves. This means that we could not have the gfx
backend do the zooming on these coordinates.

Manage them as a separate parameter that is passed down to the gfx
backend, so we can perform zoom at any precision. This results in much
less glitches when playing with GL at high resolutions.