gfx/gl: pass poly information as vertex attributes
The current way to render using GL uses uniforms to pass poly-wide
parameters, which requires one draw call per poly. This is quite
inefficient.

Instead, duplicate the poly parameters once per vertex, and pass them as
vertex attributes. Used in conjunction with the primitive restart
feature of indexed rendering, this allows us to render hundreds of polys
in a single draw call and is much more efficient.

An even more efficient way to do would involve multi draw indirect
rendering, but this is only available in later GL versions, so let's
keep this way as a default rendering method.