vm: use generic for Gfx instead of trait object
The trait object results in a virtual method call even if we statically
know the Gfx implementation that will be used. Change that in a way that
allows us to make static calls for such cases, while preserving the
ability to pass trait objects since this is our current use-case.