res: give full ownership of loaded resource data to clients
We were trying to mimic the behavior of the original game by having the
resource manager own the loaded resource's data, and the clients borrow
from it. However Rust borrow checker forced us to clone the data.

As it turns out, resources are loaded only once per scene by design ; so
it is perfectly correct (and even better) to give ownership of the
loaded data to the caller. This also simplifies the resource manager
quite a bit since it does not need to have loaded/unloaded states for
resources.