vm: move drawing of polygons into gfx
Parsing of the polygons from the segment is something that only some
backends may want to do - a shader-based renderer might prefer to parse
the polygons segment at load time into a better representation and use
that at runtime.

Make the segment be passed to the renderer as a new gfx op, and move the
polygon parsing phase into the gfx module. This requires the
introduction of a couple more traits, notably to load resources.