gfx/gl: add proper fill command for GL fill_renderer
Fill screen commands were emulated by drawing a poly that covered the
whole screen. This works, but also forces us to use u16 as the poly
bounding box type since the screen width (320) is larger that what can
fit inside a u8, which is the true type for poly bounding box
coordinates.

Adding this command will let us switch the bounding box coordinates to
u8, as they should be.